[FIX] roster may display wrong Saros support
The content provider (more specific the RosterEntryElement) is written
in such way that it only displays one presence of a user regardless of
how many presences are currently available for this specific user.

So when a user decides to login with the same account but another
presence the current Saros UI will display the Saros support for the
latest presence which in fact is misleading and may cause invitation
failures as the UI believes the contact has no Saros support.

This patch ensures that if there are multiple presences only one
presence will be used for representation and always preferring the
presence with Saros support.

Change-Id: Id148fbbe29e0ff30f15f9836af3c24a4e9815c2a
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2911
Tested-by: Jenkins CI
Reviewed-by: Denis Washington <denis@denisw.de>