[INTERNAL][I] Replace path implementation by Java7
This commit replaces the current buggy IntelliJ implementation
of IPath with the one from the server, using Java7 capabilities.
As soon as the core is switched to Java 7, this implementation
can be moved to the core.

This is the same as 2937 [1], but this time the build
infrastructure should be ready for Java wq7.

[1] http://saros-build.imp.fu-berlin.de/gerrit/#/c/2937/

Change-Id: I8c0a8b63dcf1764c30b89482337976f9562fd494
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2937
Tested-by: Jenkins CI
Reviewed-by: Denis Washington <denis@denisw.de>
Reviewed-by: Holger Schmeisky <holger.schmeisky@gmail.com>
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2988
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3040
Reviewed-by: Etienne Girard <etienne.g.girard@gmail.com>