[REFACTOR][I] Remove ChecksumCache
In IntelliJ, files may belong to more than one project (=workspace).
As such it is not possible to link a plain file to a unique workspace,
like it is done in Eclipse (where there can only be one workspace).
Without an associated workspace, IResource-objects can not exist
in Saros core.
Since the ChecksumCache is just an optimization, it is disabled in
Saros/I, in order to leave the interface intact for Saros/E.

Change-Id: Id1d109a11ba54003c8d02c382a7aa41c69674b3f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2925
Tested-by: Jenkins CI
Reviewed-by: Holger Schmeisky <holger.schmeisky@gmail.com>