[INTERNAL][HTML] Adds abstract BrowserPage to reduce code duplication
First this renames the interface BrowserPage to IBrowserPage in order to
follow our code convention. Furthermore all BrowserPages share the same
PATH="html/dist/", and have common implementation. This common behavior
is now encapsulated in the abstract BrowserPage. BrowserFunctions and
Renderer are now injected on instantiation of each Page using this super
class.Renderer and BrowserFunction are stored in a list instead of
recreating them on each call of "getJavascriptFunctions()".

This Change also include the adjustment of some naming for  methods and
fields in the BrowserPages in order to better reflect their purpose.
WEB-PAGE is changed to HTML_DOC_NAME and getWebpage to
getWebpageResource.

Due to a circle reference between Page->ClosePageBrowserFunction->Page,
each pages still holds an public static field with their HTML_DOC_NAME.
This field is needed in the DialogManger in order to close them.

Change-Id: Ieea3013f68b31e7b3568c8a81fee6f40d8a5a3be
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2949
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>