[API] Move LineRange to core
Eclipse's LineRange has been replaced with a core implementation. This
was done in preparation for moving the RemoteEditorManager to the core.

This class originates from Saros/I and holds simple values only. Saros/E
needs a small subset of the plugin impl of LineRange only and the new
LineRange can take its place.

Pair-programmed-with: Franz Zieris
Change-Id: I9d399575b1a3fe12d20c787402d41059c202d066
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2946
Tested-by: Jenkins CI
Reviewed-by: Denis Washington <denis@denisw.de>