[API] Move TextSelection to core
Eclipse's TextSelection has been replaced with a core implementation.
This was done in preparation for moving the RemoteEditorManager to the
core.

This class originates from Saros/I and holds simple values only. Saros/E
needs a small subset of the plugin impl of TextSelection only and the
new TextSelection can take its place.

Change-Id: I8765688fe32afdcac2f3ee60d4ecc694a911b433
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2947
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Denis Washington <denis@denisw.de>