[API] Move SarosCoreContextFactory to core
SarosCoreContextFactory exists in both Eclipse and IntelliJ and was
meant to be moved to the core. However there are some minor problems
with that.

I took the Saros/E variant and moved it to the core. The following
components were removed:
   * MUCNegotiationManager (Eclipse only)
   * SkypeManager (Eclipse only)
   * IsInconsistentObservable (not part of the core yet)
   * AwarenessInformationCollector (not part of the UI yet)

The above components are now part of SarosEclipseContextFactory and/or
SarosIntelliJContextFactory.

Change-Id: I88df3a17cb06cb21b8ff5f3a67b0081186fb28c4
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2971
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Denis Washington <denis@denisw.de>