[API][HTML] Introduces JavaScriptAPI to trigger JS events
This introduces a central place for JS API calls. Previously the JS
functions where directly called inside the renderers and
browser_functions classes. It was not clear
what function are provided by JS to the JAVA developer without knowing
where to look at in the UI.frontend for these functions. Also
documentation was missing about the usage and the purpose of these
functions.

All JS calls are now done via this class, that abstracts the
actual called scripts for the Java side.

This also eliminates the dependencies to the GSON library in all UI
classes. This make it easier to change the serialization of our data
model if we desire to, as there is only JavaScriptApiUtils to adapted
for such a change. For the same reasons, changes in the actual JS API
are easier to reflected on Java side.

Change-Id: I87a013d69b26c514a21778ce48e615b7b5472f36
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2780
Tested-by: Jenkins CI
Reviewed-by: Nina Weber <nina.weber@fu-berlin.de>
Reviewed-by: Christian Cikryt <c.cikryt@googlemail.com>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>