[INTERNAL][HTML] Adjusts JS functions to match backend
Currently the invitation wizard don't show up in the HTML UI. This is
caused by the previous rework of the browser functions package, which
included a renaming of the function "showStartSessionWizard" to
"showSessionWizard". The JS methods are now adjusted to match the
backend counterpart again.

Furthermore, the sendInvitation() browser function is added to the
wizard and will be executed when hitting the "finish" button. 
 
Change-Id: I9aab5cd113e0ef15f018ae9bf12f186e60b3d419
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2998
Tested-by: Jenkins CI
Reviewed-by: Nina Weber <nina.weber@fu-berlin.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>