[INTERNAL] change field injection via SarosPluginContext
Currently any error during field injection is just swallowed and just
logged leaving the object to initialize in an unknown state which leads
to further error at a later point which is more than hard to debug.

This patch ensures that the application will fail early.

In addition the ChildContainer and ChildContainerProvider classes were
deleted and the relevant code was directly moved to the initComponent
method of the SarosContext class so that it is more understandable how
the injection works for components that are not available in the
application context scope.


Change-Id: I56fa47c3cbf98724bd248abd02d6e4a8a49c77c5
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3000
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>