[INTERNAL] redesigned SarosContext class
The context class now is responsible by itself to instantiate all
objects. Before this change this was done by a call to
SarosContext#getComponents(). As it is not very obvious that this method
must be called although the return value is never used, the logic is now
replaced with an initialize method.

This patch also removes several methods that became obsolete.

Change-Id: I7180ed7ea3373fce19f60554e735c067927d62a5
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3004
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>