[INTERNAL] make HTMLUI work again
SarosPluginContext#initComponent was never designed to exploit
annotation injection. In fact we do support annotation injection in our
PicoContainer setup, but these will only work well when using a NOP CTOR
or a CTOR that does not rely on injected variables.

Again this SarosPluginContext#initComponent was desgined for objects
that are created after the context and need several references and NOT
to prettify variable declarations. If annotation injection do not work
with your component -> use CTOR injection. If you got a CTOR with 
> 10 paraments -> think about the current design.

Change-Id: I9b70b7b4487898061e7a3559f0068b959f98f70a
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3017
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>