[API] Create AbstractSarosLifecycle
In this patch the abstract class AbstractSarosLifecycle has been created
and put into the core. It abstracts the creation and disposal of
components from Saros Context Factories.

Each Saros platform (e.g. Saros/E and Saros/I) creates a
CoreContextFactory plus some additional factories. Since the logic for
this is duplicated across platforms, the introduction of
AbstractSarosLifecycle adds a more generic and common way of doing this.

This is one change in a group of three (see related changes).

Change-Id: Ic176fe14e7e72220828657502c6ed1bf5e5f8e46
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2982
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>