[API] Extend AbstractSarosLifecycle (Eclipse)
This patch extends and implements the abstract class
AbstractSarosLifecycle for the Eclipse platform.

The new EclipseSarosLifecycle adds SarosEclipseContextFactory,
HTMLUIContextFactory (if activated), and more to the Saros context.
Additionally the ConnectionHandler and ISarosSessionManager had to be
moved into the lifecycle class. They are
needed for the context factory to function.

This is one change in a group of three (see related changes).

Change-Id: I8ad9b4405bb407a13a0129bb23b7d65ec9e1654c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/2984
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>