[HTML][REFACTOR] Removes the term 'Web' from the UI project
While the HTML UI architecture uses web technologies (HTML, Javascript),
there are no online operations involved. The term 'web', as in the
method 'getWebpageResource()', therefore adds no value but just makes
the names longer. 

*'getWebpageResource' is renamed to 'getRelativePath' as this is more
precise in what this method does. 
*renames the 'webpages' package to 'pages'
*renames IWebResourceLocator to IUiResourceLocator
 
In addition the naming of some variables and JDoc in the DialogManager
are improved. This is manly to  to avoid confusion about the
usage of the page resource as an identifier in the mapping.

Change-Id: I67640a367e34d48f25c3b16f54f714fb58e8c668
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3024
Reviewed-by: David Sungaila <david.sungaila@arcor.de>
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>