[FIX][I] Fix hangup during project negotiation
This patch fixes a hangup when trying to cancel the project negotiation
on the client side. It can be tested by:
- Sharing the module succesfully a first time
- Modifying a file and setting it read-only so that project negotiation
  fails
- Sharing the same module again

A modal window progress indicator is displayed during the project
negotiation until the negotiation is finished. When cancelling the
negotiation some code was queued synchronously in IntelliJ dispatch
thread using modality state NON_MODAL. This means that the execution is
suspended until there is no modal displayed, and causes a deadlock.

In the process, modified the message displayed on an error during
project negotiation so that it does not show the jid of the remote user
if the negotiation was cancelled locally.

Change-Id: If79f9e78ff0e58142f1ddd3b1a0d441bf2bde348
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3044
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>