[HTML][REFACTOR] Introduces a new folder structure
Restructured the HTML GUI Frontend to an feature-oriented directory
structure. This helps (especially backend-) developers to find the
specific files which need to be changed in order to see the effect of
e.g. a model-change or a new function which has to be invoked by s.th.
from the frontend, as highly related files can now be found in the same
directories.

Updated the build-process accordingly -> changed the path to the
"app.js" and made "Browserify" build the .jade-files, because those are
now in more than one subdirectory and the old module which did that job
wasn't suitable anymore. Added the possibility to
"watch" for file changes and build automatically, which improves the
workflow of the frontend-developer a lot.

Also updated all package dependencies and modified the CI Test according
to the path changes.

Change-Id: I61b1bdfa0b42b54715b2377cc3f753cccb9edb51
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3063
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>