[INTERNAL] move observable classes to dedicated package
The Session and Project negotiation observable are part of the
SessionManager. Therefore they were moved and made package private and
also removed from the context.

This patch also re-enables one API call in the SessionManager.

There is also a minor change on session negotiation cancellation. The
session is no longer terminated but stays active and so has to be closed
either manually or offers the possibility to invite the same or another
person again.

Change-Id: I2b25bd5fcbf701dd7cbc5f2c76cd4d48be34b7a9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3029
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>