[INTERNAL][HTML] Reorganize ContextFactory
Since the context factory is constantly growing, it becomes harder to
read.

In order to easily find the place to add new classes (components), this
patch introduce a dedicated function for every kind of created
components. The pseudo categorization for browser functions are removed
as it was sometimes a bit clunky to decide where to put a new function.
Instead it now follows a strict alphabetic order.

Footnote: The todos are also removed, as they don't have anything to do
with the context factory itself.

Change-Id: I87f8db3369ccda51ec42fd88158c6f03f67dda17
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3060
Tested-by: Jenkins CI
Reviewed-by: David Sungaila <david.sungaila@arcor.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>