[API][HMTL] Browserfunctions are available in every Browser
Currently we limiting the available functions in each html
browser/dialog by only injecting a subset of all available
Browserfunctions(BF).

This have no beneficial, but makes adding and changing
dialogs harder for UI developers. Furthermore this creates an
unnecessary dependencies between a Page and BrowserFunction. This in
turn cause circle dependencies when a BF needs a Page instance to
operate.

To solve these problems the BrowserCreator now inject every bf in every
created browser. This is done via the new BrowserFunctions class, which
holds a list of all available Browserfunctions.

The page package is changed accordingly.

Change-Id: I9032003a162d45af73e24801ea627ccbb918a38f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3033
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>