[UI][HTML] Display all contacts in StartSessionWizard
Displays all contacts in the StartSessionWizard, instead of only currently
online contacts. Offline Contacts are disabled (grey and not clickable),
so they can't be invited to a session.

Previously it wasn't clear why there was only a blank screen and no
contacts to chose from, when no one was online.

Change-Id: Id189e8b8096b98c4d50a7a0dc252cc5e21b55e3b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3071
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>