[REFACTOR][HTML] Change name of AmpersandWizard/Wizard to WizardView
The name AmpersandWizard implied that it is part of the
ampersand.js-framework while it is in fact self-written. The name
"wizard", which I temporarily introduced in change 3063 didn't make it
clear enough what it actually is.

"WizardView" instead informs the Developer, that this is the
wizard-template and can be used as a extendable View (just as
AmpersandView, which is used widely across the whole project).

Change-Id: I978e1b4843c2b25eac9470129b5c9ffe24c2d7e0
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3072
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>