[API] Distinguish between ACTIVE and PASSIVE consumers
The EditorManager overwrites the IActivityConsumer.exec() method to
dispatch incoming activities on its own to two (one, after this patch)
other consumers. The reason is that the EditorManager relies on those
other consumers (RemoteEditorManager is one of them) to be up-to-date
regarding a certain activity, before it starts acting on the activity by
itself.

In the process of extracting the FollowMode from the EditorManager (see
subsequent patches), there would have been another class (the yet to
come FollowModeManager) that relies on the RemoteEditorManager to be
up-to-date.
So instead of dealing with such dependencies explicitly, this patch
divides all ActivityConsumers into two types: Passive consumers which
only record the current state, and active consumers which actually
perform changes based on the incoming activities and the states recorded
by the passive consumers.
Naturally, the passive consumers are notified first, then come the
actives.

By default, all existing consumers were treated as active, with the
notable exception of the RemoteEditorManager which is the first passive
consumer. Others are likely to follow, but not in this patch.

This patch also update the JTourBus tours that come along the changed
IActivityConsumer interface to reflect those changes.

Change-Id: I484460a6f5173dffc1cf3edc5f83402f6db7a0ab
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3111
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>