[API] Extract FollowModeManager from EditorManager
The EditorManager deals with many different responsibilities. This patch
extracts the FollowMode feature into a separate component. Since the
Saros/I EditorManager is very unclean, this patch mostly affects the
Saros/E side and does not yet extract the FollowMode in Saros/I.

The new FollowModeManager allows IFollowModeListener to hook in and get
notified about changes to the FollowMode status and the various reasons
it might get stopped automatically.

The FollowModeManager is session-scoped, as the feature only makes sense
when a Session is running.

There are many small task entangled with this larger task:
- Deprecate ISharedEditorListener.followModeChanged(), and use the new
  IFollowModeListener instead.
- Pull up a few public "utility" methods from EditorManager to the
  interface IEditorManager: openEditor(), closeEditor(),
  adjustViewport(), and jumpToUser()
- Un-deprecation and type-safety for SarosSession.getComponent(). Being
  constantly warned by the Java compiler does help in making sure the
  references are properly reset to allow the GC to do its job.

Change-Id: I3181d8464e1b605a510d53d42ee3a958620dca70
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3113
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>