[INTERNAL][E] Inline private method in EditorManager
There is also a theoretical change in the user experience: If the "Jump
to user" feature is used in a state where the viewport of the remote
user's editor is not known, that editor is no longer opened, as there is
nowhere to jump anyway.
This is "theoretical" because it's unknown whether such state where even
possible.

Change-Id: I2b41fedaabe8d5496b6ab982a17abdf6b1c2dc4d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3127
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>