[API][UI] Remove the AccountStoreFacade and Account model
Since XMPPAccounts are already serializable, there is no need to
translate them into another Account model class. Therefore the UI API is
simplified to use serialized XMPPAccount as parameters instead of plain
text. Therefore the the AccountStore can now be directly used without
the need of a facade.

As a result the Account model and AccountStoreFacade are removed. 

Change-Id: Iafd33d3c506bdab293e7a255180a1856e0e768ef
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3175
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>