[INTERNAL][HTML] Switch from Jade to Handlebars
Since the Template Engine "Jade" has a rather high learning curve, it is
very complicated for new developers to change even just a minor thing
(especially without any experience in frontend development).

In contrast, "handlebars.js" has a standard HTML-syntax with only minor
additions. Since this is way more intuitive, it is much easier to learn
and therefore it is a better alternative compared to "Jade".

This commit exchanges all jade-files for their handlebars equivalent and
updates the corresponding file references in all js-files.

Change-Id: I678773c18b29e286514951ddfee21c1b6f4a3e36
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3188
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>