[INTERNAL][CORE] reimplemented IBBStreamService
Currently there is an abstract transport class that tries to reduce code
bloat by implementing "shared" code that can be used by various stream
services.

Problems:

1. There are only 2 implementations since years.
2. The Socks5 implementation overwrites alot of there code and so
3. The Socks5 implementation is hardly maintainable

This patch moves code from the skeleton and place it directly in the IBB
implementation. As this may increase code bloat it is still easier to
maintain.

Change-Id: Ie4427957eecadae9b590b28b2a04b79a4ae1a69b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3206
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>