[INTERNAL][CORE] disable SO_LINGER option
It is not working well. Beside the wrong timeout (HINT: always read
carefully the Javadoc !) it is causing TCP RST packets to be sent
instead of TCP FIN packets.

This results in connections not closed correctly (even it does not harm
as it is detected anyway).

For now let the OS handle the closure of TCP connections.

Details: The issue is likely to happen if you close the connection while
there is still pending data to be read (Socket#shutDownInput is not a
solution as it is up to the TCP stack implementation, and the Windows
TCP stack does NOT what the Javadoc states, see:
http://stackoverflow.com/questions/19201451/discarding-input-from-socket
)


Change-Id: I41a44572bf32879a315d4573a5094f1768bc7876
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3212
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>