[FIX][I] NullPointerException at the start of a Saros/I-Session
The ConsistencyWatchdogServer used the function
IEditorManager.getRemotelyOpenEditors() to get the SPath of all remote
editors. The function was being called before the session was completely
initialized. This causes a NullPointerException because the
UserEditorStateManager is still null at that point.

To circumvent this problem, the function
UserEditorStateManager.getRemotelyOpenEditors() checks whether the
UserEditorStateManager is null and returns an empty set in such cases.

Change-Id: Idf60e649e7c22472ee1e1145aabb38416f0a98ff
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3215
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>