[NOP][HTML] Ran npm format & lint
After reintroducing the auto-formatting and code-linting in commit 5b760d1
this commit integrates the (propsed) changes of both scripts into the
existing code. From now on, the scripts should be run before every commit
to maintain a well-formatted and clean JS-code-base.

Because the linting proposed many changes which were not desired, some
more rules and exceptions were added to "jshint".

Change-Id: I83d5d22a63ed050ea2a7bf1ba82ec40676eab045
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3221
Tested-by: Jenkins CI
Reviewed-by: Matthias Bohnstedt <matthias.bohnstedt@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>