[FIX][I] #872 Fix synchronization issues
The synchronization in Saros/I was only working on the editor that was
active during the session creation. This was caused by the different
editor-path all being mapped onto the active editor.

This patch fixes the issue by using the
projectAPI#openEditor(VirtualFile) function to get each editor and then
map them correctly.

After all Editors are opened, the initially selected editors are made
active again. This does currently not work reliably with a split editor
view containing duplicate files.

Change-Id: I8f53b033254ed3e0d2726924ce6a7631837884ba
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3240
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>