[API] Remove getRemotelyOpenEditors from IEditorManager API
IEditorManager#getRemotelyOpenEditors is a function returning all
editors opened by a client. A UserEditorStateManager object is used to
implement this functionality by calling
UserEditorStateManager#getOpenEditors. This is somewhat problematic
because the IEditorManager (and its implementations) are application
bound while the UserEditorStateManager is session bound. This allows
calls to the UserEditorStateManager while there is no session present
which means that the UserEditorStateManager is null.

This issue was initially fixed by checking in the implementations of
IEditorManager#getRemotelyOpenEditor if the userEditorStateManager is
already initialized (and thereby not null) and returning an empty set
otherwise.

This change is a more general fix which circumvents the issue by
replacing all calls of IEditorManager#getRemotelyOpenEditors.

This is relatively easy because there is only one usage of function. It
is in WatchdogConsistencyServer#calculateChecksums. The call is replaced
by passing a userEditorStateManager to the constructor of the
ConsistencyWatchdogServer and then calling
UserEditorStateManager#getOpenEditors directly.

After this change there were no further usages of the function
getRemotelyOpenEditors. This allows the removal of the API function
IEditorManager#getRemotelyOpenEditors and its implementations.

To clean up the API, the method IEditorManager#getLocallyOpenEditors was
renamed to #getOpenEditors and the in the implementations used
variable locallyOpenEditors was also renamed to openEditorPaths to avoid
confusion with the method #openEditor.
This was done to remove any language suggesting that the API could also
handle remote editors.

Change-Id: I2e8e57ce37044bcc7748e925d139aa82df5aeaaa
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3217
Tested-by: Jenkins CI
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>