[INTERNAL][I] Change implementation of ProjectAPI#openEditor()
The implementation of ProjectAPI#openEditor(VirtualFile) was based on
creating a new editor, making it active and then returning the editor
returned by the call FileEditorManager#getActiveTextEditor().

This implementation now uses FileEditorManager#openTextEditor(...) to
create a new editor and directly return it. This also allows the
addition of the active-flag to the method, which determines whether the
newly opened editor should be selected as the active editor or not.

The methods LocalEditorHandler#openEditor(...) and
LocalEditorManipulator#openEditor(...) subsequently have the active-flag
added to them.

Change-Id: I17ca8b1f666541d3702e7d8abdeb5b32744fdf91
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3241
Tested-by: Jenkins CI
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>