[INTERNAL][UI] changes FileTransfer ETA algorithm
This patch changes the algorithm for calculating the ETA of the file
transfer.

It now uses a sample interval of one second to calculate the ETA instead
of calculating the ETA over the transfered bytes so far.

This change should result in a not so "jumpy" ETA.

TODO: average the values to avoid spikes, see:
http://saros-build.imp.fu-berlin.de/gerrit/#/c/3080/

Change-Id: Ic89b5db9d54e2dc8a5a243a766adb34bde949f50
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3249
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>