[INTERNAL][I] Use project objects held by session
The FileSystemChangedListener was always creating new projects while
adjusting to local changes. This was problematic as incoming activities
are executed using the project objects held by the session, which were
not correctly updated.

To solve this issue, a method was added that scans the pool of available
projects, looking for the project the passed resource belongs to.

This change makes the method
IntelliJWorkspaceImpl#getProjectForPath(...) obsolete and subsequently
removes it. This also removes the need to set the workspace in the
FileSystemChangeListener.

Change-Id: Ia004d5cee0e17bfb643aedb114241888a4307bfc
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3251
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>