[API][CORE] changes SPath#getResource behavior
This method no longer performs IO / VFS structure queries but always
returns a IResource instance regardless if it exists.

See also bug: https://sourceforge.net/p/dpp/bugs/448/

It adds no benefit if this method would return null because it hides
issues as long as #getFile and #getFolder will return non null while
#getResource would return null.

In addition #getFile and #getFolder will not throw an exception if the
resource is not of the given type.

Change-Id: I849b28be652b94b7a6a8e0361fc4fdccf8273ba6
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3255
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>