[CORE][API] adds methods to FileListDiff
The current way those FileListDiff stuff works is impossible for project
newcommers to understand nor to analyse. There are so many "hidden"
assertions in the logic. This patch adds some methods that are currently
not used. They will be integrated with another patch.

E.g:

you create a custom FileList A and B. A contains file foo/bar.txt, B
contains bar/foo.txt

If you perform a diff, the added and removed folder set is EMPTY.

The class contains a set of methods that mutate the state of an current
instance (seems to be partial sharing stuff). The diff instance should
be immutable.

Change-Id: I32e81d2bc0215a5ad355023424d787bdbffedff3
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3284
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>