[FIX][CORE] #793 Partial Sharing overwrite non shared files
This patch ensures that it is no longer possible to overwrite files in
partial sharing mode where the folder is not part of the sharing.

This fix is very unelegant as it will log errors instead of preventing
the logic to log such errors.

Clearly not the best solution but sufficient for now, as partial sharing
is not really used at all.

TODO: The FileListDiff logic still needs to be changed to ensure that
such folders are properly synchronized which is not the case since
years.

E.G

Alice setup:

folder: foo/bar

Bobs setup:

file: foo/bar/foo.txt

If you share folder foo/bar, Bobs file foo.txt is marked as shared and
will result in an inconsistency if this file is opened in an editor.


Change-Id: I563b176a4774572e32fa93b4f307e8f7fe2c7c12
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3301
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>