[REFACTOR][E] Remove indirection in EditorManager
The EditorManager had an ISharedEditorListener which reacted on
editorActived() events -- but only if the origin of said event was a
remote user.
The only way ISharedEditorListener.editorActivated() is called with such
parameters, is by EditorManager.execEditorActivity() itself, where we
already know that the origin of the event is a remote user.
There is simply no reason for such indirection.

Change-Id: I9b9dcade29e9fc5fc0709736db1888fa83c00706
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3361
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>