[FIX] Notify AccountStoreListeners after setAccountFile()
Without this trigger, the HTML GUI is not provided with the list of
accounts.
The alternative approach of providing the responsible AccountRenderer
with an initial list of account in AccountRenderer.ctor() or
AccountRenderer.addBrowser() does not work well with local STF tests,
as at the point where these two methods are called, the AccountStore
still contains the global (i.e. developer's) accounts, and not yet
the STF test accounts.

Change-Id: I47406073e7f7bd4c92d5f66c7caee3ab1ff30278
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/3362
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>