[STF][API] Introduce IRemoteHTMLView to abstract technical details
The HTML GUI supports multiple browser widgets to display different
parts of the GUI.
A UI designer might decide to put a certain dialog in a separate browser
widget or to render it as an overlay in the main view. For a tester,
these two cases should not be conceptually different.

The new IRemoteHTMLView interface introduces exactly this abstraction:
The decoupling of browser widgets from conceptual parts of the GUI to
interact with.

Change-Id: I95cdcadb37f4e31ce653f329bc41594af7ec2c28
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3367
Tested-by: Jenkins CI
Reviewed-by: Jannis Fey <jannisfey@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>