[INTERNAL][HTML] Relax cyclic deps, delete 'BrowserFunctions' helper
The sole purpose of the 'BrowserFunctions' class was to break cyclic
dependencies of the kind:
 - CloseAccountWizard -> DialogManager
 - DialogManager (implemented by EclipseDialogManager)
 - EclipseDialogManager -> BrowserCreator
 - BrowserCreator -> List<JavascriptFunction>
   (which includes CloseAccountWizard)

The last step in the above chain is not necessary for initializing a
BrowserCreator instance. The list of all JavascriptFunctions is only
needed when a new browser instance should be created by the
BrowserCreator.
This patch therefore adds a BrowserCreator.addJavascriptFunction()
which is called by every browser function's start() method, i.e. when
the PicoContainer created all objects.

Change-Id: Icc92cc9d4d5b4999b7c851355cb41d8212c1a5a4
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3373
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>