[FEATURE][I] Add hook to exchange module type information
Add ModuleTypeNegotiationHook which allows the exchange of module type
information during the session negotiation. This information is used
to notify the participants if the type of a shared module is not known
to one of the clients.

Standard module types are: JavaModuleType, WebModuleType,
UnknownModuleType and EmptyModuleType. They can be found in the
package com.intellij.openapi.module. Other module types can be added
through IntelliJ IDEA plugins.

Change-Id: I52ca8018a6c3fe4727937fff3a599c8cf7f4dbb2
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3394
Tested-by: Jenkins CI
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>