[INTERNAL][I] Fix LocalEditorManipulator#openEditor()
Add check whether the resource exists and is shared to
LocalEditorManipulator#openEditor(...) as this would otherwise lead to
error trying to add non-shared resources to the EditorPool.

Use IntelliJProjectImplV2#findVirtualFile() as it uses IntelliJ
OpenAPI to get the VirtualFile. Furthermore, it does so while avoiding
being dependent on absolute paths like
ResourceConverter#toVirtualFile().

Add log messages on success and failure of the method.

Change-Id: I60dd888be4e75f1e9580d125cc79e001a9c434e3
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3413
Tested-by: Jenkins CI
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>