[INTERNAL][I] Fix VirtualFile lookup in #closeEditor()
Use IntelliJProjectImplV2#findVirtualFile() as it uses the IntelliJ
OpenAPI to get the VirtualFile. Furthermore, it does so while avoiding
being dependent on absolute paths like
ResourceConverter#toVirtualFile().

Remove editor from EditorPool even if the VirtualFile does not exist
or could not be found.

Add log messages on success and failure of the method.

Change-Id: I02968d82c21d415b87eec4d5a44f938eac502003
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3416
Tested-by: Jenkins CI
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>