[INTERNAL][I] Add method to refresh module objects
This method is needed as the creation of new modules as part of the
session negotiation includes the creation of a stub module, which
later on has to be reloaded with the correct module file.

This functionality is not currently present but will be added with one
of the next changes.

Adds an exception which is thrown if the module reload fails as I
could not find a fitting already existent exception.

Change-Id: Id6aefb2c15f252ce6da4c4d0e359f3296172e921
Reviewed-on: https://saros-build.imp.fu-berlin.de/gerrit/3423
Tested-by: Jenkins CI
Reviewed-by: Tobias Bouschen <tobias.bouschen@googlemail.com>